+++ server/src/main/java/com/linbit/linstor/api/pojo/QuerySizeInfoRequestPojo.java
@@ -1,0 +1,24 @@
+package com.linbit.linstor.api.pojo;
+
+public class QuerySizeInfoRequestPojo
+{
+
+    private final String rscGrpName;
+    private final AutoSelectFilterPojo autoSelectFilterData;
+
+    public QuerySizeInfoRequestPojo(String rscGrpNameRef, AutoSelectFilterPojo autoSelectFilterDataRef)
+    {
+        rscGrpName = rscGrpNameRef;
+        autoSelectFilterData = autoSelectFilterDataRef;
+    }
+
+    public AutoSelectFilterPojo getAutoSelectFilterData()
+    {
+        return autoSelectFilterData;
+    }
+
+    public String getRscGrpName()
+    {
+        return rscGrpName;
+    }
+}